generate default source
palmed the default source generation off to the Source class. Doesn't make sense for there to be a constructor for the Source class that creates a default because presumably the header files will be instantiating these things and we don't want to be overwriting existing matrices when we set its innards point,unit.