Generate default shape
Learning. I can overload methods, but I shouldn't overload constructors, or have any functionality with constructors that would be taking input. We need a method to handle taking input. This way all class-scope variables can be initialised ahead of time.

So you will see the addSoource and addShape methods are default constructors. I think this should be ok. but then i have the generateDefault which is the manual constructor, when we are considering putting in inputs.