Begin coding tracestep
since i am appending to distances instead of insertCol, inf should come from shape. So while I create fillInf, I dont need to use it just yet.

Placed generate inside append so that we dont need to call generate at the high level, it is now automatic.

i export numrays from rays to build the distances array but since it is a vector and we automate the generate, we dont need to assign size ahead of time. So maybe i dont need numrays to be public just yet.

I add a begin and end function to shapes so that i can perform range based looping