filled gaps, reset function, reshape
generate for rows and cols ensures that the matrix is empty and reset by creating zero columns first.
appending columns should update numCols. might be more instances of this in the future.
built reset function which updates the value of the matrix itself so that we have void output. saves repeating steps at the end of each function.
Built reshape which takes in rows and cols now.