normCol
just working through createNewBasis in Rays. had to build normCol function and to achieve this i needed a sum function and replace col to enable building col as a matrix so that i could utilise the sum function as a member of the matrix class, as well as the multiply function which i think is the most convenient way even though i have to insert it back into the matrix with replaceCol.