rotate90
Rx and Ry are private functions because I cannot set a matrix variable in the matrix class (infinite loop). I notice perhaps many of my public functions could be private. I will have to clean this up next.

Rotate works on only 1 vector luckily. Created supporting function: isXaxis to test the vector.

i realised a bug in reset where the generate step works on rows and cols input it created a matrix of zeros. appending to that would just make a matrix twice as wide. So i make it create rows  with 0 cols and append to that to get the right size. bingo.