shape COB and traceLowRes
Working through scene functions. Next up write a high level traceDistance to break up the problem into chunks.
For this push I learnt pointsCOB for a shape was a MatrixList because for each ray we have a COB that applies to all points of the shape. Each matrix in the MatrixList represents a single ray. so we create a temporary matrix while assembling this, to make it a little easier to digest when reading the code.
I notice that multiply function does not automatically generate a the matrix that is being outputted. Perhaps we could put this step within the multiply function? For now, I have put this step in the rays COB as well as shape COB functions.