distanceLinePlane
Since we are using a double matrix to do indexing, there are a few steps such as converting to int so i create a function for this - indexPoint. it also simplifies inputs into this function for this task. along those lines distanceLInePLane only needs to take in rays as an argument rather than the depth of speciifc inputs. this is thanks to the class variables that shape already possesses.
subtract method solidifies the now routine format of having a matrix function with void output but then two inputs which come from elsewhere. output matrix generation takes place as the function is called, but initialisation takes place beforehand.