triangle interior. How to catch edge?
finished this and its internal functions.
I tried to simplify the triangle interior function by providing functions for each exit. Most of the space is taken by what i now call constructions, which greatly reduced the space.

check polarity returns true if the ray misses but the other checks return false. not sure if this is confusing, but it allowed me to void an exclamation mark.

One big issue is if the rays go between triangles. I need to fix this in the future. There are two places to catch this. The first is the trianglePointsFlat for AQB in checkPolarity as it indicates that Q lies on the flat edge of the triangle. The second is if compareAngles == 0.

I hesitate implenting the feature yet because it is likely that there will be two intersections, one for this triangle and one for the adjacent triangle. These functions listed above are currently not setup to give a flag that the edge has been touched.
Then it will need to be handled elsewhere to choose which direction to go (flags are likely to appear in pairs), and which rays are actually the same ray. Alternatively we could just trace both and we will get a higher ray density.. These are the options to explore later on. It would make sense to choose one of the two rays. but this should be a check case after tracing all triangles. So I should store it in a special place in the shape class (assuming different shapes cannot be tangent to each other) so at the end of tracing the shape, only then do we re-process these intersections in the new scope. This will reduce any inefficiencies in the normal running of the trace.
If the second triangle it intersects is side on to the ray, still ignore that intersection. But this will happen automatically if i change nothing.