2nd half of trace step
Wrote the inputs to rays.update and i prepare the inputs outside of the function so that i dont have to parse in the normals and the distances matrices which are local to the scene class, as well as j, but perhaps this could remove some of the fluff in tracestep, especially if we are doing more preparation within rays.update already.

Still deciding how to deal with rays that dont hit the shape or the triangle. Next time might be an appropriate time for this consideration. At the moment i place zeros in distances and normals to keep the matrix sizes consistent with numrays for looping later.

I don't believe i need matrixlist includded in shapes header file but correct me later on if there is a reason this was here.

Next: write rays.update()

On an optimisation note, Currently we trace every shape in the line of sight of a ray completely, then we choose which is the closest shape, and update the ray based on the results from that shape. However for many shapes in the line of sight, we are calculating the triangle normal of each intersection before we have decided whether we are even tracing this shape. It will be tempting to optimise this based on the complexity of trianglenNormal() however the current solution is the most elegant since the scope of the triangles is encapsulated in the shape loop and the scope of shapes is encapsulated in the scene scope.