updateLocation, reflect, refract
Predominantly equations.

I'm not satisfied with the elegance of the equations in terms of their readability. Partly this is because I cannot remember where the equations come from. Perhaps I can comment this better once I look back at the wikipedia page for refraction and reflection in vector form. Partly also because I don't have the best method for evaluating an equation without breaking it into multiple lines and instantiating a bunch of random innards and labelling them. Perhaps the reader may find it makes sense after reading the equation in comment form. I'm mainly referring to refract, and costhetaRefract - whatever that is. It needs to be named better. Where does this come from? I want to comment that in.