Proposed Solution to Circular Dependencies
While I haven't solved it here, i will outline my proposed diagnosis of all of my errors and the proposed solution.

Circular dependencies exist everywhere within my program because for example matrix functions return matrixlist types but matrixlist requires matrix to be included for its definition. It is my understanding that the compiler spots these circular dependencies and doesn't completely declare or include all of the classes before they are used, giving me errors on compilation.

The elegant solution to this is to use Forward Declarations in the header files, exclusively, and then in the implementation files, include all of the relevant header files, at all depths of the tree.

However this solution alone would lead to errors because for a function to return a class type that has been declared without a full definition, it does not know how much memory to allocate because it does not know its size or structure. So, what I propose to remedy the situation is to return pointers to the forward-declared types, rather than returning the objects directly.

Some things to be aware of while doing this. I need to create a name for my pointers. I need to ensure the full definintion is available when the pointers are dereferenced in the implementation file. Using the 'new' operator will dynamically allocate memory for the object on the heap.

And, importantly, I need to deallocate the memory when the object is no longer needed with the delete command, exercising proper memory management.

I plan on restructuring all of my functions with this in mind. It will be a big job.

First I need to understand good memory management practice and when it is a good time to deallocate the memory