circuluar dependencies
for classes that appeared only as arguments, i could use object references and then forward declare, with inclusion in the implementation files.

All the inclusions we are left with are the types that appear as outputs to functions. Classes that contain multiple of each singular class (for example as Shapes is to Shape), I have included the singular class in the header file for these only. The only other locations for inclusions are in implementation files where the classes are used. Hopefully that should be enough to solve the problem of circular dependencies.