debugging
issue in append() was that a numcols was being added to in every row.

fixed for row and column vectors although note when it says 1 col that may only be true for some rows during construction. could lead to future erros. I won't be proofing it for now, i can debug if it comes to it.