debugging dot() and append()
dot() was producing 0s and it was because i was iterating through rows of a row vector.

append was creating a new col for an existing matrix that contained 0 cols. When append(inMat) called append(i,val) this would pick up on the 0 cols and give it 1 col, and then append(inMat) would add a col to it. The case for 0 cols hadn't been considered yet, and i believe this is because i now hadn't inteded the matrix to be generated before append is called, with recent changes. So the way to append now is to crack on and do it, with no need to generate. So I removed generate() in cross() and now append works in this case.