Fix append numcols issue
append assigns numcols to 1 if the matrix is has 0 cols. so if appending a matrix to a matrix, and it is already constructed but has 0 cols, it assigns a cols to 1 then adds numcols from inmat to this. so there are one too many cols if appending a matrix to a 0 col matrix.

solution was to swap iteration of cols with rows and add a column every new iteration, however adding it before the append(i,v) is called will mean that it will not encounter the cols==0 case.