fixed traceLowRes Issue
tracelowRes was returning false every time. after further inspection of the x,y vals it seems the minRow function was not producing the minimum. Indeed the minRow function was using the max element vector function. whoops.

Something I have learnt in this process is that if the ray sits on the edge of the rectangle made by the points, it counts as a reject. Worth noting for the future in case we want edge rays to count as an intersection.