debugging triangle interior
didn't initialise k to 0, ABC wasn't constructing.

indexPointCOB had a segmenation fault when accessing connectivity. First of all there was no indication in the comments what k is, so when connectivty (2,3) is indexed with (2,0) and there is no third row, questions are asked! particularly of its index which is (i,j) which wouldn't make sense given i is ray, triangle is j and connectivity is triangle point by triangle.
Looking at this immediately I can see that we need to change the index for connectivity to:
triangle j : row
triangle point k : col
then i needed to comment what k is to make it clearer.
Additionally, we need to be indexing ray i (not k) from pointsCOB as this is its highest dimension.