inspecting triangle interior
I believe that check polarity will not work since it just measures if the points are above or below the x axis. I will come up with a more micromanaged version to be sure it will work. by shifting the left most point to 0,0 I can test the other points for being above or below the x axis for polarity.