New triangle interior - edge case intention
After further thought I realise I can perform a much simpler triangle interior algorithm. It uses barycentric coordinate system. Should be simpler to understand and faster to execute.

As for the edge case, the intention is to send off a true, but raise a flag somewhere so that when we have duplicate intersections of the same ray into two different triangles, we can reject one.