debugged triangleInterior
I was having troubles with inverse. Solved by slicing 3D vectors into 2D vectors so that the matrix is 2x2 then also testing if the determinant is 0 before calculating the inverse. I couldn't catch the runtime error in my debugging terminal so maybe I don't see those while debugging, then I also count it as a false interior because it means that either AB or BC is 0 which means B is A or B is C and therefore the triangle as seen from the ray is just a straight line, so it will not intersect. I have not considered other edge cases.