debug dot product
dot product was designed for column vectors but it iterated internally through columns so clearly we had a problem. This was solved by using the agnostic get function (with a single input) and choosing rows or cols to go through on the from the first vector. if it is a 1x1 vector it still works as it cycles over the single col.