debugging traceDistance
I wasn't squaring the components of the vector I was finding the magnitude of.

Problem for next time. The line plane intersection was giving a negative distance because the normal vector was pointing downwards, towards the incoming ray. I'm not sure how I can standardise the normal to point in the direction of the ray so that we can reject negative distances that signify a triangle that lies behind the ray location.

my intended solution is to find the projection of the normal onto the ray and see if it is negative. So if i take the dot product of the unit vector of the ray with the normal, then i get the shadow of the normal over the ray. if the dot product is negative, flip the normal.