triangle interior is wrong
triangle interior equation is wrong.
for i=1,j=1 scenario we know Q should be within {A,B,C} but we are testing whether Q is in {0,AB,AC} and since Q is (0,-1), it comes up as external to the triangle, which is wrong.

I need to re-write the final checks for interior for a correct algorithm.