potential fix to triangle interior algorithm
We were using the Q from the original axes but now we are using the Q from point A (as B and C are now also from point A) so then Q should be a linear combination of positive scalars of AB and AC because it is in the right spot.

not tested.