fixed triangle interior
after much sorting through the code i realise that the issue was with the initial creation of the default shape.

There were two mistakes.
First, the number of points was wrong.
Second, the connectivity matrix had a 2 placed in the first entry that should have been a zero.

So the algorithm was fine. Very embarrassing.