fix bug in append()
I was getting a crash at an unknown location. Stepping through the program I narrowed it down to the zeros function and the append function.
Zeros was outdated and not setting constructed variable to true therefore i updated it to reflect the correct way of making matrices - generate. which already sets it to zeros by default, doing its job.

Second was the append() function which was meant to add a column onto the end but was only adding a column the length of a row, which in the case of adding zeroes to a column vector of let's say 3 by 1, it was adding a single 0 on the end but then the matrix was claiming there are 2 columns. The bottom right entry and the one above it are then completely unknown and causing some kind of undefined error as they were not doubles.