plotting works
To test the functionality of the ray trace I have decided to visualise using python's matplotlib pyplot quiver function which is very straightforward.

This works and plots correctly.

One thing to note is that the length of the unit vectors feeds the length of the vectors in the matplotlib plot.

therefore the distancesAcc vector can be consumed within the cpp code and multiplied to the unit vectors.

Also worth noting to fix for next time, the unit vectors are not actually normalised.