draw shape
Can now draw shapes as vectors by adding them to the list of vectors for the python to plot.

realised i was indexing points a few times so i collected it into a function.
I would index COB as well but it's separate.

tested. works. however the distance of the 3rd ray is too short. the 6th ray is in the correct spot though. so it finds the right distance, it's not making its way into the output matrix