downloaded assimp via vcpkg
i changed all references that were encountered through the debug process that came up with error of file/directory not found for references to header files in the assimp folder. for example i changed <assimp/types.h> to "types.h". Not sure if this was an exclusive change but i kept on going until the debugging errors stopped. Note I still have an error.