i wrote an stl file
turns out writing to an stl file is easy. The following format was simply renamed as a .stl then it was easily opened using my windows 3d viewer (i assume this is windows 10 default program)
assuming that the stl file we want to import is in ascii then i should be able to import an stl file no worries. Although i cannot assume this.

At least for now i can experiment with the stl type as an output for the trace, to utilise the wide range of tools that can visualise STLs.