add triangles for edge
triangle points have already been added now the connectivity matrix describes what points are connected through triangles. the algorithm i have decided to conduct these connections is drawn on paper as the best explanation of what is happening and so i am troubled as to how to make that easy to read and understand in the code without visualisation aids.

Either way next step will be to break up this massive function into smaller parts which may assist in understanding what is happening later down the track.