tested ray polygons via python plot
Test failed because of a segmentation error when referencing connectivity matrix.

I've noticed the connectivity matrix default structure is to run each triangle down the columns, which is actually against the convection of all other matrices that append new items as columns.

So I will change the default shape connectivity structure to columns, and then get back to testing this.