modify inclusions in assimp library
my lack of cpp knowledge means im not sure why they assume each header file is included as if it were a part of the standard library. Leading me to question if i downloaded/installed the library properly. For now I have renamed them as normal inclusions and hope that only header files are necessary for the library to work.