Create python RayTraceGUI
Decided to use python as a GUI, which can interface with the cpp code in numerous ways.
Started with the basics, building the createSource commands in a frame. There will be a plot in the side window to demonstrate the 2D results to the user using pyplot.

interfacing with cpp could be via creating a python library