createSource pop-out window finished
I have all the data I require for the input, now the next step is to bind this data and call the c++ library from python.

Actually before I do this I will want to build shape inputs, so that I have enough to call the cpp trace.

For now I should preview the source in the create Source tab in a pyplot figure. Then it updates when the user makes modifications to the input data.