re-arranged createSource code
Re-arranged the layout and also changed the method for containers so that it is heirarchichal and relies on pack instead of grid so that it is easier to edit. it is also simpler to visualise as there are 3 main frame and every label entry pair has a frame. only potential source of confusion is the multitude of frame references for each item and linking them all up can be tedious.