fixY works. add createSource
Intended to save the data for the rectangle when we hide the yval from the user's view in circle mode. so that when we go back to rectangle mode, the data is still there.

added CreateSource button in the middle

Next step is to get real data into the plots.
A. build scatter plot demo to replicate what we expect from cpp algorithm
B. When changing fields in the GUI, the plots update (goes for both plots)
C. build divergence line plot