duplicate cpp within python
in order to get the quick feedback from within tkinter callbacks as the user modifies their selection I have decided it is more concise to replicate what the cpp does in python so that calling the cpp library can be done in a concise way, i.e. it takes all of its inputs and builds the scene in one go, not having to store that data before we call the trace function.
only downside (other than a potential delay) is the need to replicate the cpp function every time we create a new source creation method - essentially duplicating the code, which is generally considered a bad idea considering making future changes in two locations instead of one.