scaling axes correctly
aspect ratio is set to equal.
in addition, the xlim and ylim are set just outside of the source bounds.
i was getting issues when there was an empty entry because the min and max values were nonsense so i decided instead of plotting a rectangle with dimension 0 during the entry being edited, i simply dont plot it until there is a value in the entry