python can now call cpp!
using subprocess python module, i have just attached it to the plot button in the gui as a test. and to measure my results i made it so i save my results in text files and the time created for the text files matches when i pressed plot. This means I can now use the GUI to call the cpp algorithm, I believe the python will be the most convenient and agile way to create the sources and objects from user input, then convert them to the basic types that cpp can interpret so that we are not stuck with the cumbersome task of filling data into cpp from within cpp. It will inherit all of its data from python defaults. This should simplify the defaults process within cpp and make all inputs as uploads.