begin
I decided to put the read source line within the sources file because we cannot choose a specific line to read, rather we just read the next line. If i make a source function to read a source then it starts from the beginning, i would be reading already read code which is inefficient.

I'm having trouble building the first matrix column with my current functions, so i think i will build some matrix functions to handle this, rather than handling it in the sources file, when it will need to be replicated elsewhere too.