rotate points
Rotate points in the direction of v to generate a beam source in the direction of v. then we need to save this data into the text file remembering to broadcast v then once tested with a 3d plot we can work on building functionality for circles and then point sources.