automatically set plot limits
rather than having to set manually as before.
I realise that a more general approach to inputting string data into a float is a catch all so i implement a try except structure for the data input for x,y.

for density it wanted an integer, but this should be a float so we can go below 1.

Got errors in the scatterPlot where an error in width height or density would not stop the plotting from using their values. The spooky issue i have now is that the function not continue if it doesn't interpret these values, but it still continues.. it shouldnt print "but still in scatter" because we have left scatter with the return.