scale vectors to limits
now that i am potentially shrinking or expanding the limits of graph to meet the location of the source rays, i would also like to shrink and expand the length of the vectors to match the limits since these vectors don't have any intrinsic length yet.

However this does not work yet since I need to scale the x y and z to the maximum along any dimension so that i maintain 1to1 scale.

Also I will be separating the plot commands into a separate function and plot source calls this and all it does is grab the x y  z which will be self data.