fix origin bug and append error
I was appending numpy arrays to python lists which was giving size errors so I use a numpy array instead and its inbuilt append method.

Next error was that when trying to rotate points I would be trying to rotate the origin (0,0,0) which would give issue when determing the cos of the angle towards it which involves a divide by zero. So to remove this issue I note when the normal for that point is 0 and disregard its rotation by replacing its normal with 1 to prevent a creation of NaNs then replace the points back as the origin after the process produces nonsense values for those points.